feat: Complete UWB integration and wheel loader system enhancements
🚀 Major Features Added:

## UWB (Ultra-Wideband) Integration for EKF2
- ✅ Added NoopLoop LinkTrack UWB driver with full PX4 integration
- ✅ Implemented complete UWB fusion pipeline in EKF2
- ✅ Added UWB range measurements for indoor positioning
- ✅ Created comprehensive UWB parameter system (16 parameters)
- ✅ Added UWB quality checks and signal validation
- ✅ Integrated UWB data flow: Driver → sensor_uwb → EKF2 fusion

### UWB Driver Features:
- Split architecture: .hpp/.cpp files following PX4 conventions
- LinkTrack protocol support with enhanced diagnostics
- Signal quality metrics (RSSI, LOS confidence, multipath detection)
- Configurable anchor positions and sensor mounting
- Robust error handling and parameter validation

### EKF2 UWB Fusion:
- Added UWB aid source with innovation gating
- Range-based position estimation with anchor mapping
- Adaptive measurement noise based on signal quality
- Comprehensive parameter system for tuning
- Build system integration with CONFIG_EKF2_UWB

## Wheel Loader System Enhancements:
- 🔧 Enhanced articulated chassis control systems
- 🔧 Improved traction control and slip estimation
- 🔧 Advanced terrain adaptation algorithms
- 🔧 Load-aware torque distribution
- 🔧 Updated message definitions for load sensing

## Build System & Configuration:
- ✅ Added UWB Kconfig options for proper feature selection
- ✅ Updated board configuration (hkust_nxt-dual-wl-rear) with UWB support
- ✅ Fixed parameter naming conventions (≤16 chars)
- ✅ Corrected parameter units following PX4 standards
- ✅ Added ekf2_timestamps logging for UWB debugging

## Technical Improvements:
- Fixed double promotion warnings in UWB signal processing
- Implemented proper aid source status management
- Added UWB buffer system with time synchronization
- Created comprehensive parameter validation
- Enhanced CMakeLists.txt structure for conditional compilation

📊 Validation Status:
- ✅ Successful compilation on ARM Cortex-M7 target
- ✅ Memory usage: 54.49% Flash, 10.90% SRAM
- ✅ All UWB files compiled and linked successfully
- ✅ Firmware generation completed: hkust_nxt-dual-wl-rear_default.px4

🔗 Integration Pipeline:
NoopLoop Driver → sensor_uwb topic → EKF2::UpdateUwbSample() → setUwbData() → UWB Buffer → controlUwbFusion() → Position Estimation

Ready for runtime testing and validation! 🎯

+++ src/modules/ekf2/EKF/aid_sources/uwb/uwb_checks.cpp
@@ -31,7 +31,7 @@
  *
  ****************************************************************************/
 
-#include "uwb_checks.hpp"
+#include "aid_sources/uwb/uwb_checks.hpp"
 #include <math.h>
 
 using namespace estimator;
@@ -181,7 +181,7 @@ bool UwbChecks::runMultipathCheck(const uwbSample &uwb)
 		const float frequency_hz = 6.5e9f; // Center frequency for UWB
 		const float c = 299792458.0f; // Speed of light
 		const float wavelength = c / frequency_hz;
-		const float expected_path_loss_db = 20.0f * log10f((4.0f * M_PI * uwb.range_m) / wavelength);
+		const float expected_path_loss_db = 20.0f * log10f((4.0f * static_cast<float>(M_PI) * uwb.range_m) / wavelength);
 
 		// Assuming 0 dBm transmit power and some margin for multipath
 		const float expected_rssi = -expected_path_loss_db - 10.0f; // 10dB margin
